+++ src/main/java/com/modagbul/BE/domain/fire/entity/Fire.java
@@ -1,7 +1,5 @@
 package com.modagbul.BE.domain.fire.entity;
 
-import com.modagbul.BE.domain.mission.entity.Mission;
-import com.modagbul.BE.domain.team.entity.Team;
 import com.modagbul.BE.domain.user.entity.User;
 import com.modagbul.BE.domain.usermission.entity.UserMission;
