+++ src/main/java/com/modagbul/BE/domain/mission/domain/repository/MissionRepository.java
@@ -1,8 +1,8 @@
-package com.modagbul.BE.domain.mission.repository;
+package com.modagbul.BE.domain.mission.domain.repository;
 
-import com.modagbul.BE.domain.mission.dto.MissionDetailDto;
-import com.modagbul.BE.domain.mission.dto.MissionListDto;
-import com.modagbul.BE.domain.mission.entity.Mission;
+import com.modagbul.BE.domain.mission.application.dto.MissionDetailDto;
+import com.modagbul.BE.domain.mission.application.dto.MissionListDto;
+import com.modagbul.BE.domain.mission.domain.entity.Mission;
 import com.modagbul.BE.domain.usermission.constant.Status;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
