+++ src/main/java/com/modagbul/BE/domain/mission/domain/service/MissionService.java
@@ -1,12 +1,12 @@
-package com.modagbul.BE.domain.mission.service;
+package com.modagbul.BE.domain.mission.domain.service;
 
-import com.modagbul.BE.domain.mission.Exception.InvalidDueToDate;
-import com.modagbul.BE.domain.mission.Exception.MissionAuthDeniedException;
-import com.modagbul.BE.domain.mission.Exception.NotFoundMissionException;
-import com.modagbul.BE.domain.mission.dto.MissionListDto;
+import com.modagbul.BE.domain.mission.exception.InvalidDueToDate;
+import com.modagbul.BE.domain.mission.exception.MissionAuthDeniedException;
+import com.modagbul.BE.domain.mission.exception.NotFoundMissionException;
+import com.modagbul.BE.domain.mission.application.dto.MissionListDto;
 
-import com.modagbul.BE.domain.mission.entity.Mission;
-import com.modagbul.BE.domain.mission.repository.MissionRepository;
+import com.modagbul.BE.domain.mission.domain.entity.Mission;
+import com.modagbul.BE.domain.mission.domain.repository.MissionRepository;
 import com.modagbul.BE.domain.team.entity.Team;
 import com.modagbul.BE.domain.team.repository.TeamRepository;
 import com.modagbul.BE.domain.team_member.repository.TeamMemberRepository;
@@ -26,7 +26,7 @@ import java.util.Date;
 import java.util.Iterator;
 import java.util.List;
 
-import static com.modagbul.BE.domain.mission.dto.MissionDto.*;
+import static com.modagbul.BE.domain.mission.application.dto.MissionDto.*;
 
 @Service
 @RequiredArgsConstructor
