+++ src/main/java/com/modagbul/BE/domain/mission/exception/InvalidDueToDate.java
@@ -1,6 +1,6 @@
-package com.modagbul.BE.domain.mission.Exception;
+package com.modagbul.BE.domain.mission.exception;
 
-import com.modagbul.BE.domain.mission.constant.MissionConstant.MissionExceptionList;
+import com.modagbul.BE.domain.mission.application.constant.MissionConstant.MissionExceptionList;
 
 public class InvalidDueToDate extends MissionException{
     public InvalidDueToDate() {
