+++ src/main/java/com/modagbul/BE/domain/mission/exception/MissionAuthDeniedException.java
@@ -1,6 +1,6 @@
-package com.modagbul.BE.domain.mission.Exception;
+package com.modagbul.BE.domain.mission.exception;
 
-import static com.modagbul.BE.domain.mission.constant.MissionConstant.MissionExceptionList.MISSION_AUTH_DENIED;
+import static com.modagbul.BE.domain.mission.application.constant.MissionConstant.MissionExceptionList.MISSION_AUTH_DENIED;
 
 public class MissionAuthDeniedException extends MissionException {
     public MissionAuthDeniedException() {
